+++ CHANGELOG.rst
@@ -10,8 +10,10 @@ Unreleased
 
 Added
 ~~~~~~
+- better support for data from terrestrial laser scanners
 - has_original_id for datasets. Returns true if all pointclouds have original_id
 - PointCloud.from_file now supports timestamp input or "from_file"
+- diff with "nearest" to calculate distance to nearest point from another pointcloud
 
 Changed
 ~~~~~~